deps: bump spdlog to 1.15.3 (fmt 11) + migrate fmt usage
GitHub moved the `windows-latest` CI image to Visual Studio 2026, whose
MSVC STL removed the legacy non-standard `stdext::checked_array_iterator`.
spdlog 1.8.5's bundled fmt 7.1.3 still used it on its Debug `_SECURE_SCL`
path, breaking only the Windows-Debug build (Release uses the plain-pointer
path). Our code didn't change — the compiler under it did.

Bump deps/spdlog 1.8.5 -> 1.15.3 (bundled fmt 7.1.3 -> 11.2.0), which
removed that path entirely, and migrate shards' fmt usage for fmt 11:

- Custom formatters: const-qualify every `format()` (fmt 10+ calls it on a
  const formatter) and drop the throwing `parse()` (broke fmt 11's consteval
  format-string check); covers ops_internal, gfx/fmt, gfx/shader/fmt and
  render_graph_builder. Also fix a latent missing return in LayoutPath.
- Add formatters for types fmt 11 no longer formats implicitly, placed with
  their definitions so they are always visible: FastString (fast_string.hpp),
  gfx::UniqueId (unique_id.hpp), WGPUTextureFormat (gfx_wgpu.hpp).
- Wrap runtime/ternary format strings in `fmt::runtime()` (exception.hpp,
  gfx error helpers, trait.cpp, math_shards.cpp).
- `#include <spdlog/fmt/std.h>` for std::thread::id formatting.
- Call-site fixes preserving prior output: SHTrait cast, boost stacktrace
  to_string, boost::filesystem::path .string(), http verb to_string, and
  enum->int (BuiltinTarget) to match fmt 7's implicit enum rendering.

Verified: clean Release build and full local test suite (367 pass; the one
failure is llm.shs needing the `hf` CLI which isn't installed locally).

+++ shards/modules/gfx/gltf.cpp
@@ -64,7 +64,7 @@ static OwnedVar getGltfBuiltinTargetPath(animation::BuiltinTarget target) {
     pathStr[1] = 't';
     break;
   default:
-    SHLOG_ERROR("Ignoring builtin target: {}", target);
+    SHLOG_ERROR("Ignoring builtin target: {}", (int)target);
     break;
   }
   pathStr[2] = 0;
@@ -374,7 +374,7 @@ struct GLTFShard {
       node->trs.translation = toVec<float3>(value);
       break;
     default:
-      SHLOG_WARNING("Ignoring builtin target: {}", target);
+      SHLOG_WARNING("Ignoring builtin target: {}", (int)target);
       break;
     }
     node->update();
